+++ NEWS.md
@@ -2,7 +2,9 @@
 
 ## Changes in version 0.1.1
 
-- ...
+- The run function now looks if there is an updated release and proposes to
+  install it. Also, it provides a menu to select tutorials, in case no arguments
+  are provided.
 
 
 ## Version 0.1.0